🎛 notebook UI controls
