+++ packages/site/src/components/ContentBlocks.tsx
@@ -3,7 +3,11 @@ import { SourceFileKind } from 'myst-common';
 import type { GenericParent } from 'myst-common';
 import { useNodeRenderers } from '@myst-theme/providers';
 import classNames from 'classnames';
-import { ClearNotebookCell, RunNotebookCell } from './ComputeControls';
+import {
+  ClearNotebookCell,
+  RunNotebookCell,
+  RunNotebookCellSpinnerOnly,
+} from './controls/NotebookCellControls';
 import { useIsAComputableCell } from '@myst-theme/jupyter';
 
 function isACodeCell(node: GenericParent) {
@@ -29,7 +33,6 @@ function Block({
   className?: string;
 }) {
   const renderers = useNodeRenderers() ?? DEFAULT_RENDERERS;
-  const { computable, ready } = useIsAComputableCell(id);
   const children = useParse(node, renderers);
   const subGrid = 'article-grid article-subgrid-gap col-screen';
   const dataClassName = typeof node.data?.class === 'string' ? node.data?.class : undefined;
@@ -40,17 +43,24 @@ function Block({
     <div
       key={id}
       id={id}
-      className={classNames('relative group/block border border-black', className, dataClassName, {
+      className={classNames('relative group/block', className, dataClassName, {
         [subGrid]: !noSubGrid,
       })}
     >
       {pageKind === SourceFileKind.Notebook && isACodeCell(node) && (
-        <div className="hidden sticky top-[80px] z-20 opacity-70 group-hover/block:opacity-100 group-hover/block:flex">
-          <div className="absolute top-0 -right-[28px] flex md:flex-col">
-            <RunNotebookCell id={id} />
-            <ClearNotebookCell id={id} />
+        <>
+          <div className="flex sticky top-[80px] z-20 opacity-70 group-hover/block:opacity-100 group-hover/block:hidden">
+            <div className="absolute top-0 -right-[28px] flex md:flex-col">
+              <RunNotebookCellSpinnerOnly id={id} />
+            </div>
           </div>
-        </div>
+          <div className="hidden sticky top-[80px] z-20 opacity-70 group-hover/block:opacity-100 group-hover/block:flex">
+            <div className="absolute top-0 -right-[28px] flex md:flex-col">
+              <RunNotebookCell id={id} />
+              <ClearNotebookCell id={id} />
+            </div>
+          </div>
+        </>
       )}
       {children}
     </div>
